[Refactor] #120 - ScanView-CameraManager 리펙토링 완료
