+++ PennyPack/Utilities/Double+Ext.swift
@@ -5,8 +5,4 @@ extension Double {
         let divisor = pow(10.0, Double(places))
         return (self * divisor).rounded() / divisor
     }
-    
-    func asString(withDecimalPlaces places: Int) -> String {
-        return String(format: "%.\(places)f", self)
-    }
 }
